upgrade K8s clusters to v1.11.0
This attempts to follow best practices from
https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/kubeadm-upgrade-1-11/.

kubeadm upgrades cannot skip a minor version so upgrading to 1.10.5 must
be completed as the first step before upgrading to 1.11.0.

CoreDNS is not included because of SELinux incompatibilities introduced
in https://github.com/kubernetes/kubernetes/pull/64473/files.

This does not update Replicated to a version compatible with K8s
v1.11.0. The kubernetes-upgrade.sh script should probably be merged with
the kubernetes-init.sh script to accomplish that.

Untested on airgap.