Fix k8s cluster upgrades
Skip the recommended kubectl drain step. This is probably recommended to
prevent downtime when upgrading a cluster with enough nodes to run the
workload while each node is upgraded invidually, but in practice the
extra load of rebalancing pods and restarting Rook was found to cause
failures and longer downtime.

Fix the missing kubeadm-flags.env file on remote nodes. In kubeadm 1.11,
this file is how the kubelet is configured to use CNI plugins and run
with the correct cgroup driver.