more insecure registry configuration
When upgrading an airgap cluster to ha, the replicated pod can be
scheduled on a worker node, but that node would not have been configured
to allow insecure pulls from the in-cluster docker registry. This adds a
prompt to tell users to manually configure docker on worker nodes before
proceeding with the ha upgrade.

For the master node where the configuration can be updated in the
script, prefer to use /etc/docker/daemon.json since it can be reloaded
with SIGHUP rather than required a full restart.